renaming of tests and _simple_dataset for clarity and consistency
