+++ fedbiomed/common/dataset/_image_label_dataset.py
