+++ fedbiomed/common/dataset/_mappings.py
@@ -22,12 +22,12 @@ from fedbiomed.common.dataset_controller import (
 from fedbiomed.common.exceptions import FedbiomedError
 
 from ._custom_dataset import CustomDataset
-from ._medical_folder_dataset import MedicalFolderDataset
-from ._simple_dataset import (
+from ._image_label_dataset import (
     ImageFolderDataset,
     MedNistDataset,
     MnistDataset,
 )
+from ._medical_folder_dataset import MedicalFolderDataset
 from ._tabular_dataset import TabularDataset
 
 DATASET_CLASSES_PER_TYPE = {
